move generic_pricing_model module
